+++ Router/src/RestaurantContainer.test.jsx
@@ -2,29 +2,29 @@ import React from 'react';
 
 import { render } from '@testing-library/react';
 
-import { useDispatch, useSelector } from 'react-redux';
-import RestaurantPage from './RestaurantPage';
+import { useSelector, useDispatch } from 'react-redux';
+import RestaurantContainer from './RestaurantContainer';
 
 jest.mock('react-redux');
-describe('RestaurantPage', () => {
+describe('RestaurantContainer', () => {
   beforeEach(() => {
     const dispatch = jest.fn();
     useDispatch.mockImplementation(() => dispatch);
-    useSelector.mockImplementation((selector) => selector({
+    useSelector.mockImplementation((selector) => selector(({
       restaurant: {
         id: 1,
         name: '마법사 주방',
+        address: '서울시 강남구',
       },
-    }));
+    })));
   });
 
-  it('renders name', () => {
-    const params = { id: '1' };
-
+  it('renders name and address', () => {
     const { container } = render((
-      <RestaurantPage params={params} />
+      <RestaurantContainer restaurantId="1" />
     ));
 
     expect(container).toHaveTextContent('마법사 주방');
+    expect(container).toHaveTextContent('서울시');
   });
 });
